# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/pm-tasks": {
      "get": {
        "operationId": "PmTasksController_getAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PM Tasks"
        ]
      },
      "post": {
        "operationId": "PmTasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePmTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PM Tasks"
        ]
      }
    },
    "/api/v1/pm-tasks/{id}": {
      "get": {
        "operationId": "PmTasksController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PM Tasks"
        ]
      },
      "put": {
        "operationId": "PmTasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PM Tasks"
        ]
      },
      "delete": {
        "operationId": "PmTasksController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PM Tasks"
        ]
      }
    },
    "/api/v1/templates": {
      "get": {
        "operationId": "TemplatesController_getAll",
        "parameters": [
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Templates"
        ]
      },
      "post": {
        "operationId": "TemplatesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Templates"
        ]
      },
      "delete": {
        "operationId": "TemplatesController_deleteByNameAndDept",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Templates"
        ]
      }
    },
    "/api/v1/templates/{id}": {
      "delete": {
        "operationId": "TemplatesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "CMMS API",
    "description": "The Computerized Maintenance Management System (CMMS) API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ChecklistItemDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "requiresPhoto": {
            "type": "boolean"
          }
        },
        "required": [
          "text"
        ]
      },
      "CreatePmTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "frequency": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "nextDueDate": {
            "type": "string"
          },
          "estimatedHours": {
            "type": "number",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "In Progress",
              "Pending Approval",
              "Done",
              "Overdue"
            ]
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemDto"
            }
          },
          "partsRequired": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "partsUsed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recordNotes": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "frequency",
          "assetId",
          "productId",
          "department",
          "nextDueDate",
          "estimatedHours"
        ]
      },
...
    }
  }
}